requirements.txt
apache-airflow==2.7.2
apache-airflow-providers-postgres==5.11.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0
# Excluding problematic packages:
# google-re2 - requires C++ build tools and Abseil headers
# python-nvd3 - optional visualization package
# unicodecsv - optional package 